Migrate packages.yaml to ansible configuratoin
Instead of installing packages from a suspicious YAML file, use ansible
for cleaner, more cross-compatible functionality down the road.

As of the moment, scripts.d/ has not been implemented into packageman3
as it is planned to create its ansible version.

As of the moment, TODO: Create a method to undo ansible packages